+++ python/cudf/cudf/_libxx/__init__.py
@@ -4,6 +4,7 @@ from . import (
     copying,
     null_mask,
     rolling,
+    search,
     sort,
     stream_compaction,
     table,